Refactor transformation names
Rename transformation functions to address linter warnings.
Rename models to `Old` models for clarity
