+++ concepts/handlers_test.go
@@ -34,7 +34,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -49,7 +49,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/Dummy/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -64,7 +64,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/financial-instruments/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "FinancialInstrument"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "FinancialInstrument"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -79,7 +79,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/FinancialInstrument/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "FinancialInstrument"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "FinancialInstrument"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -106,7 +106,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/dummies/%s", "99999"), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -121,7 +121,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return nil, errors.New("TEST failing to WRITE")
@@ -136,7 +136,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return nil, rwapi.ConstraintOrTransactionError{}
@@ -151,7 +151,7 @@ func TestPutHandler(t *testing.T) {
 			req:  newRequest("PUT", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			mockService: &mockConceptService{
 				decodeJSON: func(decoder *json.Decoder) (interface{}, string, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "not-dummy"}, knownUUID, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "not-dummy"}, knownUUID, nil
 				},
 				write: func(thing interface{}, transID string) (interface{}, error) {
 					return ConceptChanges{}, nil
@@ -189,7 +189,7 @@ func TestGetHandler(t *testing.T) {
 			req:  newRequest("GET", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			ds: &mockConceptService{
 				read: func(uuid string, transID string) (interface{}, bool, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, true, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, true, nil
 				},
 			},
 			statusCode:  http.StatusOK,
@@ -201,7 +201,7 @@ func TestGetHandler(t *testing.T) {
 			req:  newRequest("GET", fmt.Sprintf("/Dummy/%s", knownUUID), t),
 			ds: &mockConceptService{
 				read: func(uuid string, transID string) (interface{}, bool, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, true, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Dummy"}, true, nil
 				},
 			},
 			statusCode:  http.StatusBadRequest,
@@ -213,7 +213,7 @@ func TestGetHandler(t *testing.T) {
 			req:  newRequest("GET", fmt.Sprintf("/locations/%s", knownUUID), t),
 			ds: &mockConceptService{
 				read: func(uuid string, transID string) (interface{}, bool, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Location"}, true, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Location"}, true, nil
 				},
 			},
 			statusCode:  http.StatusOK,
@@ -225,7 +225,7 @@ func TestGetHandler(t *testing.T) {
 			req:  newRequest("GET", fmt.Sprintf("/Location/%s", knownUUID), t),
 			ds: &mockConceptService{
 				read: func(uuid string, transID string) (interface{}, bool, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "Location"}, true, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "Location"}, true, nil
 				},
 			},
 			statusCode:  http.StatusBadRequest,
@@ -261,7 +261,7 @@ func TestGetHandler(t *testing.T) {
 			req:  newRequest("GET", fmt.Sprintf("/dummies/%s", knownUUID), t),
 			ds: &mockConceptService{
 				read: func(uuid string, transID string) (interface{}, bool, error) {
-					return transform.AggregatedConcept{PrefUUID: knownUUID, Type: "not-dummy"}, true, nil
+					return transform.OldAggregatedConcept{PrefUUID: knownUUID, Type: "not-dummy"}, true, nil
 				},
 			},
 			statusCode:  http.StatusBadRequest,
